tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.alpha.4.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.alpha.4</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.alpha.4</h1>
<p class="desc">Shadows with alpha components are correctly affected by globalAlpha</p>


<script>
var t = async_test("Shadows with alpha components are correctly affected by globalAlpha");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#f00'; // (work around broken Firefox globalAlpha caching)
  ctx.shadowColor = 'rgba(0, 0, 255, 0.707)';
  ctx.shadowOffsetY = 50;
  ctx.globalAlpha = 0.707;
  ctx.fillRect(0, -50, 100, 50);

  _assertPixelApprox(canvas, 50,25, 127,0,127,255, 2);
  t.done();

});
</script>
